feat: rename files to follow project standards fix: remove logs from test execution
