+++ tests/mock/stream-logger-mock.ts
