+++ tests/stubs/format-response.ts
